+++ inst/rmarkdown/lua/latex-div.lua
@@ -1,8 +1,17 @@
 --[[
      A Pandoc 2 Lua filter converting Pandoc native divs to LaTeX environments
-     Author: Romain Lesur, Christophe Dervieux, and Yihui Xie
+     Author: Romain Lesur, Christophe Dervieux, Atsushi Yasumoto and Yihui Xie
      License: Public domain
 --]]
+
+--[[
+  About the requirement:
+  * PANDOC_VERSION -> 2.1
+]]
+if (not PANDOC_VERSION) or (PANDOC_VERSION < "2.1") then
+    io.stderr:write("[WARNING] (latex-div.lua) requires at least Pandoc 2.1. Lua Filter skipped.\n")
+    return {}
+end
 
 text = require 'text'